xe_peak: Preserve precision in calculation
When calculating benchmark measurements, float types are
used and precision is lost. Use a long double to preserve
precision when calculating benchmarks.

Related-to: VLCLJ-508

+++ perf_tests/xe_peak/include/common.h
@@ -41,7 +41,7 @@ class Timer {
     void start();
 
     // Stop and return time in micro-seconds
-    float stopAndTime();
+    long double stopAndTime();
 };
 
 uint64_t roundToMultipleOf(uint64_t number, uint64_t base, uint64_t maxValue);
